Fix miles changing in assert

+++ TddAir/test/com/tddair/WhenPurchasingSeatUpgrades.java
@@ -71,7 +71,7 @@ public class WhenPurchasingSeatUpgrades {
 		setUpMember(Category.Red, 11000);
 		int purchasedUpgrades = purchaser.purchaseUpgradeWithCreditCard(1, "9999111155558888");
 		assertEquals(1, purchasedUpgrades);
-		assertEquals(new Integer(1000), purchaser.getMilesBalance());
+		assertEquals(new Integer(11000), purchaser.getMilesBalance());
 		assertEquals(new Integer(1), purchaser.getUpgrades());
 	}
